+++ beetsplug/bandcamp/_helpers.py
@@ -3,8 +3,7 @@ import itertools as it
 import operator as op
 import re
 from functools import lru_cache, partial
-from string import Template
-from typing import Any, Dict, Iterable, List, NamedTuple, Pattern, Tuple
+from typing import Any, Dict, Iterable, List, NamedTuple, Pattern
 
 from beets.autotag.hooks import AlbumInfo
 from ordered_set import OrderedSet as ordset
@@ -30,54 +29,55 @@ class MediaInfo(NamedTuple):
     description: str
 
 
-_catalognum = Template(
-    r"""(?<![]/@-])(\b
-(?!\W|LC[ ]|VA[\d ]+|[EL]P\W|[^\n.]+[ ](?:20\d{2}|VA[ \d]+)|(?i:vol|disc|number))
+CATALOGNUM_CONSTRAINT = r"""(?<![]/@-])(\b
+(?!\W|LC[ ]|VA[\d ]+|[EL]P\W|[^\n.]+[ ](?:20\d\d|VA[ \d]+)|(?i:vol|disc|number))
+{}
+\b(?!["%]))"""
+_cat_pat = CATALOGNUM_CONSTRAINT.format(
+    r"""
 (
       [A-Z][A-Z .]+\d{3}         # HANDS D300, CC ATOM 101
     | [A-Z-]{3,}\d+              # RIV4
-    # dollar signs need escaping here since the $label below will be
-    # substituted later, and we do not want to touch these two
-    | [A-Z]{2,}[A-Z.$$-]*\d{2,}  # HS11, USE202, HEY-101, LI$$INGLE025
+    | [A-Z]{2,}[A-Z.$-]*\d{2,}   # HS11, USE202, HEY-101, LI$INGLE025
     | (?<!\w\W)[A-Z.]{2,}[ ]\d+  # OBS.CUR 9
     | [A-z]+-[A-z]+[ ]?\d+       # o-ton 119
     | \w+[A-z]0\d+               # 1ØPILLS018, fa036
-    | [a-z]+(cd|lp|:)\d+         # ostgutlp45, reni:7
+    | [a-z]+(?:cd|lp|:)\d+       # ostgutlp45, reni:7
     | [A-z]+\d+-\d+              # P90-003
-    | (?i:$label[ ]?[A-Z]*\d+[A-Z]*)
 )
-( # optionally followed by
+(?: # optionally followed by
       (?<=\d\d)-?[A-Z]+  # IBM001CD (needs at least two digits before the letter)
     | [.]\d+             # ISMVA002.1
 )?
-\b(?!["%]))"""
+"""
 )
-_cat_pat = _catalognum.template
 
+LABEL_CATNUM = CATALOGNUM_CONSTRAINT.format(r"(?i:{}[ ]?[A-Z]*\d+[A-Z]*)")
 CATNUM_PAT = {
-    "with_header": re.compile(
-        r"(?:^|\s)cat[\w .]+?(?:number\b:?|:) ?(\w[^\n,]+)", re.I
-    ),
-    "start_end": re.compile(rf"((^|\n){_cat_pat}|{_cat_pat}(\n|$))", re.VERBOSE),
+    # preceded by some variation of 'Catalogue number:'
+    "header": re.compile(r"^cat[\w .]+(?:number\b:?|:) ?(\w.+)$", re.I | re.M),
+    # beginning or end of line
+    "start_end": re.compile(rf"(^{_cat_pat}|{_cat_pat}$)", re.M | re.VERBOSE),
     # enclosed by parens or square brackets, but not ending with MIX
     "delimited": re.compile(rf"(?:[\[(])(?!.*MIX){_cat_pat}(?:[])]|$)", re.VERBOSE),
-    "anywhere": re.compile(rf"(?<!,[ ])({_cat_pat}([ ]/[ ]{_cat_pat})?)", re.VERBOSE),
+    # can possibly be followed up by a second catalogue number
+    "anywhere": re.compile(rf"({_cat_pat}(\ /\ {_cat_pat})?)", re.VERBOSE),
 }
 
-_comp = re.compile
-PATTERNS: Dict[str, Pattern] = {
-    "split_artists": _comp(r", - |, | (?:[x+/-]|//|vs|and)[.]? "),
-    "meta": _comp(r'.*"@id".*', re.M),
-    "remix_or_ft": _comp(r" [\[(].*(?i:mix|edit|f(ea)?t([.]|uring)?).*"),
-    "ft": _comp(
+PATTERNS: Dict[str, Pattern[str]] = {
+    "split_artists": re.compile(r", - |, | (?:[x+/-]|//|vs|and)[.]? "),
+    "meta": re.compile(r'.*"@id".*'),
+    "ft": re.compile(
         r" *((([\[(])| )f(ea)?t([. ]|uring)(?![^()]*mix)[^]\[()]+(?(3)[]\)])) *", re.I
     ),
-    "track_alt": _comp(
+    "track_alt": re.compile(
         r"^([A-J]{1,3}[12]?\d|[AB]+(?=\W{2,}))(?:(?!-\w)[^\w(]|_)+", re.I + re.M
     ),
-    "vinyl_name": _comp(r"[1-5](?= ?(xLP|LP|x))|single|double|triple", re.I),
-    "clean_incl": _comp(r" *(\(?incl|\((inc|tracks|.*remix( |es)))([^)]+\)|.*)", re.I),
-    "tidy_eplp": _comp(r"\S*(?:Double )?(\b[EL]P\b)\S*", re.I),
+    "vinyl_name": re.compile(r"[1-5](?= ?(xLP|LP|x))|single|double|triple", re.I),
+    "clean_incl": re.compile(
+        r" *(\(?incl|\((inc|tracks|.*remix( |es)))([^)]+\)|.*)", re.I
+    ),
+    "tidy_eplp": re.compile(r"\S*(?:Double )?(\b[EL]P\b)\S*", re.I),
 }
 rm_strings = [
     "limited edition",
@@ -89,16 +89,19 @@ rm_strings = [
     r"free download|\([^()]*free(?!.*mix)[^()]*\)",
 ]
 
+_remix_pat = r"(?P<remix>((?P<remixer>[^])]+) )?\b((re)?mix|edit|bootleg)\b[^])]*)"
 # fmt: off
 CLEAN_PATTERNS = [
-    (_comp(r" -(\S)"), r" - \1"),                   # hi -bye      -> hi - bye
-    (_comp(r"(\S)- "), r"\1 - "),                   # hi- bye      -> hi - bye
-    (_comp(r"  +"), " "),                           # hi  bye      -> hi bye
-    (_comp(r"(- )?\( *"), "("),                     # hi - ( bye)  -> hi (bye)
-    (_comp(r" \)+|\)+$"), ")"),                     # hi (bye ))   -> hi (bye)
-    (_comp(r'(^|- )"([^"]+)"( \(|$)'), r"\1\2\3"),  # "bye" -> bye; hi - "bye" -> hi - bye
-    (_comp(r"- Reworked"), "(Reworked)"),           # bye - Reworked -> bye (Reworked)
-    (_comp(fr"(([\[(])|(^| ))\*?({'|'.join(rm_strings)})(?(2)[])]|( |$))", re.I), ""),
+    (re.compile(fr"(([\[(])|(^| ))\*?({'|'.join(rm_strings)})(?(2)[])]|( |$))", re.I), ""),       # noqa
+    (re.compile(r" -(\S)"), r" - \1"),                    # hi -bye          -> hi - bye
+    (re.compile(r"(\S)- "), r"\1 - "),                    # hi- bye          -> hi - bye
+    (re.compile(r"  +"), " "),                            # hi  bye          -> hi bye
+    (re.compile(r"(- )?\( *"), "("),                      # hi - ( bye)      -> hi (bye)
+    (re.compile(r" \)+|(\)+$)"), ")"),                    # hi (bye ))       -> hi (bye)
+    (re.compile(r"- Reworked"), "(Reworked)"),            # bye - Reworked   -> bye (Reworked)    # noqa
+    (re.compile(rf"(\({_remix_pat})$", re.I), r"\1)"),    # bye - (Some Mix  -> bye - (Some Mix)  # noqa
+    (re.compile(rf"- *({_remix_pat})$", re.I), r"(\1)"),  # bye - Some Mix   -> bye (Some Mix)    # noqa
+    (re.compile(r'(^|- )"([^"]+)"( \(|$)'), r"\1\2\3"),   # "bye" -> bye; hi - "bye" -> hi - bye  # noqa
 ]
 # fmt: on
 
@@ -107,15 +110,16 @@ class Helpers:
     @staticmethod
     def get_label(meta: JSONDict) -> str:
         try:
-            return meta["albumRelease"][0]["recordLabel"]["name"]
+            item = meta["albumRelease"][0]["recordLabel"]
         except (KeyError, IndexError):
-            return meta["publisher"]["name"]
+            item = meta["publisher"]
+        return item.get("name") or ""
 
     @staticmethod
     def get_vinyl_count(name: str) -> int:
         conv = {"single": 1, "double": 2, "triple": 3}
-        for match in PATTERNS["vinyl_name"].finditer(name):
-            count = match.group()
+        for m in PATTERNS["vinyl_name"].finditer(name):
+            count = m.group()
             return int(count) if count.isdigit() else conv[count.lower()]
         return 1
 
@@ -142,34 +146,38 @@ class Helpers:
     @staticmethod
     @lru_cache(maxsize=None)
     def parse_catalognum(
-        album="", disctitle="", description="", label="", tracks=None, artists=None
+        album="", disctitle="", description="", label="", artistitles=""
     ):
-        # type: (str, str, str, Tuple[str], Tuple[str]) -> str
+        # type: (str, str, str, str, str) -> str
         """Try getting the catalog number looking at text from various fields."""
         cases = [
-            (CATNUM_PAT["with_header"], description),
+            (CATNUM_PAT["header"], description),
             (CATNUM_PAT["anywhere"], disctitle),
             (CATNUM_PAT["anywhere"], album),
             (CATNUM_PAT["start_end"], description),
             (CATNUM_PAT["anywhere"], description),
         ]
         if label:
-            pat = re.compile(_catalognum.substitute(label=re.escape(label)), re.VERBOSE)
+            pat = re.compile(LABEL_CATNUM.format(re.escape(label)), re.VERBOSE)
             cases.append((pat, "\n".join((album, disctitle, description))))
 
-        tracks_str = " ".join([*(tracks or []), *(artists or [])]).lower()
-
-        def find(pat: Pattern, string: str) -> str:
-            """Return the match if it is not found in any of the track names."""
-            m = pat.search(string)
-            if m:
+        def find(pat: Pattern[str], string: str) -> str:
+            """Return the match if it is
+            * not found in any of the track artists or titles
+            * made of the label name when it has a space and is shorter than 6 chars
+            """
+            for m in pat.finditer(string):
                 catnum = m.group(1).strip()
-                if catnum.lower() not in tracks_str:
+                if catnum.lower() not in artistitles:
+                    if " " in catnum:
+                        first = catnum.split()[0].lower()
+                        if len(catnum) <= 5 and first not in label.lower():
+                            continue
                     return catnum
             return ""
 
         try:
-            return next(filter(op.truth, it.starmap(find, cases)))
+            return next(filter(None, it.starmap(find, cases)))
         except StopIteration:
             return ""
 
@@ -188,7 +196,7 @@ class Helpers:
         name = re.sub(r"^\[(.*)\]$", r"\1", name)
 
         for arg in [re.escape(arg) for arg in filter(op.truth, args)] + [
-            r"Various Artists?\b(?! \w)"
+            r"Various Artists?\b(?! [A-z])( \d+)?"
         ]:
             if not re.search(rf"\w {arg} \w", name, re.I):
                 name = re.sub(