+++ tests/test_search.py
@@ -1,16 +1,20 @@
 """Tests for searching functionality."""
 import pytest
-from beetsplug.bandcamp._search import RELEASE_PATTERNS, parse_and_sort_results
+from beetsplug.bandcamp._search import get_matches, parse_and_sort_results
 
 # simplified version of the search result HTML block
 HTML_ITEM = """
 <div class="searchresult data-search">
+https://www.recaptcha.net/recaptcha/enterprise.js
 <a href="{url}">
 search_item_type="a">
      {name}
      <span>some
      other stuff</span>
   by {artist}
+  <div class="released">
+      released {date}
+  </div>
   <div class="itemtype">
   ALBUM
   </div>
@@ -19,7 +23,7 @@ search_item_type="a">
 
 
 def make_html_item(data):
-    return HTML_ITEM.format(**data)
+    return HTML_ITEM.format(**data, date="26 November 2021")
 
 
 @pytest.fixture
@@ -35,8 +39,9 @@ def search_data():
 
 def test_search_logic(search_data):
     """Given a single matching release, the similarity should be 1."""
-    results = parse_and_sort_results(make_html_item(search_data), **search_data)
-    assert results == [{**search_data, "similarity": 1.0, "index": 1}]
+    expected_data = {**search_data, "date": "2021 November 26"}
+    results = parse_and_sort_results(make_html_item(search_data), **expected_data)
+    assert results == [{**expected_data, "similarity": 1.0, "index": 1}]
 
 
 def test_search_prioritises_best_matches(search_data):
@@ -58,7 +63,10 @@ def test_search_prioritises_best_matches(search_data):
         "similarity": 0.925,
     }
 
-    expected_results = [expected_result, other_result]
+    expected_results = [
+        {**expected_result, "date": "2021 November 26"},
+        {**other_result, "date": "2021 November 26"},
+    ]
 
     html = make_html_item(other_result) + "\n" + make_html_item(expected_result)
     results = parse_and_sort_results(
@@ -67,20 +75,20 @@ def test_search_prioritises_best_matches(search_data):
     assert results == expected_results
 
 
+# fmt: off
 @pytest.mark.parametrize(
     ("test_url", "expected_label"),
     (
-        (
-            "https://bandcamp.materiacollective.com/track/the-illusionary-dance",
-            "materiacollective",
-        ),
-        (
-            "https://finderskeepersrecords.bandcamp.com/track/illusional-frieze",
-            "finderskeepersrecords",
-        ),
+        ("https://bandcamp.materiacollective.com/track/the-illusionary-dance", "materiacollective"),  # noqa
+        ("https://finderskeepersrecords.bandcamp.com/track/illusional-frieze", "finderskeepersrecords"),  # noqa
+        ("https://compiladoaspen.bandcamp.com/track/kiss-from-a-rose", "compiladoaspen"),  # noqa
+        ("https://comtruise.bandcamp.com/track/karova-digital-bonus-3", "comtruise"),
+        ("https://bandofholyjoy.bandcamp.com/track/lost-in-the-night", "bandofholyjoy"),
+        ("https://bandcampcomp.bandcamp.com/track/everything-everything-in-birdsong-acoustic", "bandcampcomp"),  # noqa
+        ("https://bandcamp.bandcamp.com/track/warm-2", "bandcamp"),
     ),
 )
 def test_search_matches(test_url, expected_label):
-    test_matches = RELEASE_PATTERNS[-1].match(test_url)
-    assert test_matches.group("url") == test_url
-    assert test_matches.group("label") == expected_label
+    result = get_matches(test_url)
+    assert result["url"] == test_url
+    assert result["label"] == expected_label