+++ tests/test_tracks.py
@@ -70,7 +70,7 @@ def print_result(console, case, expected, result):
         ("¯\\_(ツ)_/¯", ("", "", "", "¯\\_(ツ)_/¯", "¯\\_(ツ)_/¯")),
         (
             "VIENNA (WARM UP MIX",
-            ("", "", "", "VIENNA (WARM UP MIX", "VIENNA (WARM UP MIX"),
+            ("", "", "", "VIENNA (WARM UP MIX)", "VIENNA"),
         ),
         ("MOD-R - ARE YOU", ("", "MOD-R", "", "ARE YOU", "ARE YOU")),
         ("K - The Lightning", ("", "K", "", "The Lightning", "The Lightning")),
@@ -140,9 +140,10 @@ def test_parse_catalognum_from_track_name(
         **json_track["item"],
         "position": json_track["position"],
         "name": name,
+        "name_parts": {"catalognum": initial_catalognum, "clean": name},
     }
 
-    track = Track.from_json(json_track, name, "-", initial_catalognum, "Label")
+    track = Track.from_json(json_track, "-", "Label")
     assert track.title == expected_title, print(track)
     assert track.catalognum == expected_catalognum, print(track)
 
@@ -168,6 +169,4 @@ def test_parse_catalognum_from_track_name(
     ],
 )
 def test_check_digi_only(name, expected_digi_only, expected_name):
-    track = Track(_name=name)
-    assert track.no_digi_name == expected_name
-    assert track.digi_only == expected_digi_only
+    assert Track.clean_digi_name(name) == (expected_name, expected_digi_only)